Разработка от 13.03.2023
-Стабилизация стула
