--- database/docker-compose.yml
@@ -1,21 +1,0 @@
-#Docker-compose файл для создания контейнера с Бд. Для МС project
-
-version: '1.0'
-
-services:
-  postgres-project:
-    image: postgres:10
-    container_name: 'postgres-project'
-    environment:
-      POSTGRES_DB: mission-project
-      POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: private_password
-      PGDATA: "/var/lib/postgresql/data/pgdata"
-    volumes:
-      - ./assets/init_db:/docker-entrypoint-initdb.d
-      - ./assets/db_data/postgres:/var/lib/postgresql/data
-    networks:
-      - database
-networks:
-  database:
-    driver: bridge
