+++ src/main/kotlin/ru/kyamshanov/mission/project/missionproject/controller/AdminController.kt
@@ -28,117 +28,5 @@ class AdminController @Autowired constructor(
     private val taskService: TaskService
 ) {
 
-    @PostMapping("create")
-    suspend fun registration(
-        @RequestBody(required = true) body: CreateProjectRqDto
-    ): ResponseEntity<CreateProjectRsDto> {
-        val projectModel = ProjectModel(title = body.title, description = body.description)
-        val responseModel = CreateProjectRsDto(
-            id = requireNotNull(projectCreatorService.createProject(projectModel).id) { "Saved entity has no Id" }
-        )
-        return ResponseEntity(responseModel, HttpStatus.OK)
-    }
 
-    @GetMapping("find")
-    suspend fun find(
-        @RequestParam(required = true, value = "id") id: String
-    ): ResponseEntity<CreateProjectRsDto> {
-        val responseModel = CreateProjectRsDto(
-            id = requireNotNull(projectCreatorService.getProject(id).id) { "Saved entity has no Id" }
-        )
-        return ResponseEntity(responseModel, HttpStatus.OK)
-    }
-
-    @PostMapping("attach")
-    suspend fun attachTeam(
-        @RequestBody(required = true) body: AttachTeamRqDto
-    ): ResponseEntity<Unit> {
-
-        val team = Team(body.participants.map { userId -> Participant(userId, Participant.Role.PARTICIPANT) })
-        teamService.attachTeam(body.project, team)
-        return ResponseEntity(HttpStatus.OK)
-    }
-
-    @GetMapping("team")
-    suspend fun attachTeam(
-        @RequestParam(required = true, value = "project") projectId: String
-    ): ResponseEntity<GetTeamRsDto> {
-        val response = GetTeamRsDto(projectId, teamService.getTeam(projectId).participants)
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @PostMapping("role")
-    suspend fun setRole(
-        @RequestBody(required = true) body: SetRoleRqDto
-    ): ResponseEntity<Unit> {
-        teamService.addParticipant(body.projectId, Participant(body.userId, body.role))
-        return ResponseEntity(HttpStatus.OK)
-    }
-
-    @PostMapping("stage")
-    suspend fun setStage(
-        @RequestBody(required = true) body: SetStageRqDto
-    ): ResponseEntity<Unit> {
-        stageService.setProjectStage(body.projectId, ProjectStage(body.stage))
-        return ResponseEntity(HttpStatus.OK)
-    }
-
-    @GetMapping("stage")
-    suspend fun getStage(
-        @RequestParam(required = true, value = "project") id: String
-    ): ResponseEntity<ProjectStageRsDto> {
-        val response = ProjectStageRsDto(stageService.getProjectStage(id).toDto())
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @GetMapping("stage/history")
-    suspend fun getStageHistory(
-        @RequestParam(required = true, value = "project") id: String
-    ): ResponseEntity<HistoryRsDto> {
-        val response = HistoryRsDto(stageService.getStageHistory(id).map { it.toDto() })
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @PostMapping("task/create")
-    suspend fun createTask(
-        @RequestBody(required = true) body: CreateTaskRqDto
-    ): ResponseEntity<CreateTaskRsDto> {
-        val taskModel = TaskModel(
-            projectId = body.projectId,
-            title = body.title,
-            text = body.text
-        )
-        val response = taskService.createTask(taskModel)
-            .let {
-                CreateTaskRsDto(
-                    taskId = requireNotNull(it.id) { "Saved task (title: ${body.title} has id equal null." },
-                    createdAt = it.createAt
-                )
-            }
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @GetMapping("task/get")
-    suspend fun getTask(
-        @RequestParam(required = true, value = "task") taskId: String
-    ): ResponseEntity<GetTaskRsDto> {
-        val response = taskService.getTask(taskId)
-            .let {
-                GetTaskRsDto(
-                    title = it.title,
-                    text = it.text,
-                    createdAt = it.createAt
-                )
-            }
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @GetMapping("task/get/all")
-    suspend fun getAllTasks(
-        @RequestParam(required = true, value = "project") projectId: String
-    ): ResponseEntity<GetTasksRsDto> {
-        val response = taskService.getLightTasks(projectId)
-            .map { it.id }.let { GetTasksRsDto(it) }
-        return ResponseEntity(response, HttpStatus.OK)
-    }
 }
