+++ src/main/kotlin/ru/kyamshanov/mission/project/missionproject/dto/FindProjectDto.kt
@@ -1,0 +1,8 @@
+package ru.kyamshanov.mission.project.missionproject.dto
+
+data class FindProjectRsDto(
+    val id: String,
+    val title: String,
+    val description: String,
+    val participants: List<ParticipantDto>
+)
