+++ src/main/kotlin/ru/kyamshanov/mission/project/missionproject/dto/ParticipantDto.kt
@@ -1,0 +1,23 @@
+package ru.kyamshanov.mission.project.missionproject.dto
+
+import ru.kyamshanov.mission.project.missionproject.models.Participant
+
+data class ParticipantDto(
+    val userId: String,
+    val role: Role
+) {
+
+    enum class Role {
+        PARTICIPANT, LEADER
+    }
+}
+
+internal fun Participant.toDto() = ParticipantDto(
+    userId = userId,
+    role = role.toDto()
+)
+
+private fun Participant.Role.toDto() = when (this) {
+    Participant.Role.PARTICIPANT -> ParticipantDto.Role.PARTICIPANT
+    Participant.Role.LEADER -> ParticipantDto.Role.LEADER
+}
